requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
awkward==2.8.4
awkward0==0.15.5
awkward_cpp==46
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blosc2==2.5.1
boost-histogram==1.5.2
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.1
click==8.1.8
cloudpickle==3.1.1
coffea==2025.3.0
comm==0.2.2
contourpy==1.3.0
correctionlib==2.7.0
cramjam==2.10.0
cvxpy==1.6.6
cycler==0.12.1
dask==2024.8.0
dask-awkward==2025.5.0
dask-histogram==2025.2.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
EnergyFlow==1.4.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
fsspec_xrootd==0.5.1
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hist==2.8.1
histoprint==2.6.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jetnet==0.2.5
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
llvmlite==0.43.0
locket==1.0.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mplhep==0.3.59
mplhep_data==0.0.4
mpmath==1.3.0
msgpack==1.1.1
multidict==6.4.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.4.3
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.2
numpy==1.26.4
osqp==1.0.4
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qpth==0.0.18
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.13.1
scs==3.2.7.post2
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
tables==3.9.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.2.2
torch-geometric==2.6.1
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uhi==0.5.0
uproot==5.1.2
uri-template==1.3.0
urllib3==2.5.0
vector==1.6.2
wasserstein==1.1.0
wcwidth==0.2.13
weaver-core==0.4.17
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wurlitzer==3.1.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
